[codeql] Fix mobutils lossy function result cast
GetMobHPScale and GetMobRBI are only used in one place, contain uint values in their respective arrays, and are assigned to uint8 results.

Changes return types of both of these functions to uint8 from float.

+++ src/map/grades.cpp
@@ -191,12 +191,12 @@ namespace grade
         return StatScale[rank][scale];
     }
 
-    float GetMobHPScale(uint8 rank, uint8 scale)
+    uint8 GetMobHPScale(uint8 rank, uint8 scale)
     {
         return MobHPScale[rank][scale];
     }
 
-    float GetMobRBI(uint8 rank, uint8 scale)
+    uint8 GetMobRBI(uint8 rank, uint8 scale)
     {
         return MobRBI[rank][scale];
     }
